Set block size for metadata to 1m
For offsets and errors, we use less than 1k per file, so the remaining
space of the block of 256m (default size) is wasted. Since we create one
offset file and one error file per partition per run. A lot of space
will be wasted. With this commit we change the block size to 1m the
minimum possible.

Change-Id: I394a020379d6ff12c2866ab6413e5d4bd9e0f4c4