Kafka offset beyond range: move them back into current range if parameters set
While setting up InputSplit for the mappers to process, if previous offset is no more coherent with current kafka partition offset (eg kafka recovered from an unclean state),
setting the parameter kafka.move.to.closest.offset can bring the offset back into range --
if previous offset is before the earliest offset, the offset is moved to the earliest offset.
if previous offset is behind the last offset, the offset is moved back to the last offset.

Change-Id: I4b11d06b8227a34bc08b8817d90309347a0318e4